// models/model-types.ts
export interface TechnologiesTypes {
  id?: string;
  name?: string;
  description?: string;
  logo?: string;
}

export interface ProjectsTypes {
  deployUrl?: string;
  projectId?: string;
  projectName?: string;
  projectDescription?: string;
  projectImg?: string;
  projectUrl?: string;
  githubImg?: string;
  vercelImg?: string;
}

export interface ContactTypes {
  user_name: string;
  user_email: string;
  message: string;
}